Use cleaner json config for mod building

+++ KerbalModderTools.Launch/KSP_Launcher.cs
@@ -1,13 +1,7 @@
 ﻿using KerbalModderTools.Deploy.Library;
 using KerbalModderTools.Library;
 using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace KerbalModderTools.Launch
 {
@@ -15,7 +9,7 @@ namespace KerbalModderTools.Launch
     {
         private readonly string _ksp_x64_dbg;
         private readonly Deployer _deployer;
-        private Process _process;
+        private Process? _process;
 
         public KSP_Launcher(IOptions<Constants> constants, EnvironmentLoader environment, Deployer deployer)
         {
@@ -27,7 +21,7 @@ namespace KerbalModderTools.Launch
         {
             this.KillRunningInstances();
 
-            if(_deployer.TryDeploy())
+            if (_deployer.TryDeploy())
             {
                 //Process.Start(_ksp_x64_dbg, "-popupwindow");
                 _process = Process.Start(_ksp_x64_dbg);
@@ -39,11 +33,16 @@ namespace KerbalModderTools.Launch
             }
         }
 
-        private void HandleProcessExit(object sender, EventArgs e)
+        private void HandleProcessExit(object? sender, EventArgs e)
         {
             AppDomain.CurrentDomain.ProcessExit -= this.HandleProcessExit;
 
-            if(_process.HasExited == false)
+            if (_process is null)
+            {
+                return;
+            }
+
+            if (_process.HasExited == false)
             {
                 _process.Kill();
             }
@@ -51,17 +50,22 @@ namespace KerbalModderTools.Launch
 
         private Process KillRunningInstances()
         {
-            foreach(Process process in Process.GetProcesses())
+            foreach (Process process in Process.GetProcesses())
             {
                 try
                 {
+                    if (process.MainModule is null)
+                    {
+                        continue;
+                    }
+
                     if (process.MainModule.FileName.StartsWith(_ksp_x64_dbg))
                     {
                         process.Kill();
                         process.WaitForExit();
                     }
                 }
-                catch(Exception e)
+                catch (Exception e)
                 {
                     //
                 }
